Fix debugging printf.
MFC after:	1 month
Sponsored by:	The FreeBSD Foundation


Former-commit-id: 5251ef99e54379f579f362d7561478af78bc8388

+++ sys/kern/kern_rctl.c
@@ -2866,7 +2866,7 @@ expr_stmt|;
 if|#
 directive|if
 literal|0
-block|printf("%s: pid %d (%s), %jd of %jd, will sleep for %ld ms (ratio %ld, available %ld)\n", 			   __func__, p->p_pid, p->p_comm, 			   p->p_racct->r_resources[resource], 			   rule->rr_amount, sleep_ms, sleep_ratio, available);
+block|printf("%s: pid %d (%s), %jd of %jd, will sleep for %ju ms (ratio %ju, available %jd)\n", 			   __func__, p->p_pid, p->p_comm, 			   p->p_racct->r_resources[resource], 			   rule->rr_amount, (uintmax_t)sleep_ms, 			   (uintmax_t)sleep_ratio, (intmax_t)available);
 endif|#
 directive|endif
 name|KASSERT
